+++ app/models.d.ts
@@ -1,9 +1,9 @@
 import Knex from 'knex';
 
 import { DbAdapter } from './support/DbAdapter';
-import { PubSubAdapter } from './support/PubSubAdapter';
+import PubSubAdapter from './pubsub';
 import { GONE_NAMES } from './models/user';
-import { UUID } from './support/types';
+import { Nullable, UUID } from './support/types';
 import { SessionTokenV1Store } from './models/auth-tokens';
 
 export const postgres: Knex;
@@ -11,25 +11,51 @@ export const dbAdapter: DbAdapter;
 export const PubSub: PubSubAdapter;
 
 export class User {
+  id: UUID;
   intId: number;
   username: string;
   readonly isActive: boolean;
+  type: 'user';
   setGoneStatus(status: keyof typeof GONE_NAMES): Promise<void>;
   unban(usernames: string): Promise<1>;
   unsubscribeFrom(targetUser: User): Promise<boolean>;
   getHomeFeeds(): Promise<Timeline[]>;
   getSubscriptionsWithHomeFeeds(): Promise<{ user_id: UUID; homefeed_ids: UUID[] }[]>;
+  isGroup(): false;
+  isUser(): true;
+  getManagedGroups(): Promise<Group[]>;
+  getBanIds(): Promise<UUID[]>;
 }
 
-export class Group {}
+export class Group {
+  id: UUID;
+  intId: number;
+  username: string;
+  type: 'group';
+  isGroup(): true;
+  isUser(): false;
+  getAdministrators(): Promise<User[]>;
+}
 
 export class Post {
+  id: UUID;
+  userId: UUID;
+  body: string;
   destroy(destroyedBy?: User): Promise<void>;
   removeLike(user: User): Promise<boolean>;
+  getPostedTo(): Promise<Timeline[]>;
+  onlyUsersCanSeePost(fromUsers: User[]): Promise<User[]>;
+  getGroupsPostedTo(): Promise<Group[]>;
+  getCreatedBy(): Promise<User>;
 }
 
 export class Timeline {
+  id: UUID;
+  userId: UUID;
+  getUser(): Promise<User | Group>;
   destroy(): Promise<void>;
+  isDirects(): boolean;
+  isPosts(): boolean;
 }
 
 export class Attachment {
@@ -37,6 +63,12 @@ export class Attachment {
 }
 
 export class Comment {
+  id: UUID;
+  body: string;
+  userId: Nullable<UUID>;
+  hideType: 0 | 1 | 2 | 3;
+  postId: UUID;
+  getPost(): Promise<Post>;
   removeLike(user: User): Promise<boolean>;
 }