+++ app/pubsub.ts
@@ -1,43 +1,31 @@
-import { dbAdapter } from './models';
+import { Comment, dbAdapter, Post } from './models';
 import { serializeTimeline } from './serializers/v2/timeline';
+import { PubSubAdapter } from './support/PubSubAdapter';
+import { UUID } from './support/types';
 
-export class DummyPublisher {
-  userUpdated() {}
-  postCreated() {}
-  postDestroyed() {}
-  postUpdated() {}
-  commentCreated() {}
-  commentDestroyed() {}
-  commentUpdated() {}
-  likeAdded() {}
-  likeRemoved() {}
-  postHidden() {}
-  postUnhidden() {}
-  commentLikeAdded() {}
-  commentLikeRemoved() {}
-  globalUserUpdated() {}
-  postSaved() {}
-  postUnsaved() {}
-  groupTimesUpdated() {}
+export class DummyPublisher extends PubSubAdapter {
+  constructor() {
+    super({
+      publish: () => Promise.resolve(),
+    });
+  }
 }
 
 export default class pubSub {
-  constructor(publisher) {
+  constructor(private publisher: PubSubAdapter) {}
+
+  setPublisher(publisher: PubSubAdapter) {
     this.publisher = publisher;
   }
 
-  setPublisher(publisher) {
-    this.publisher = publisher;
-  }
-
-  async updateUnreadDirects(userId) {
+  async updateUnreadDirects(userId: UUID) {
     const unreadDirectsNumber = await dbAdapter.getUnreadDirectsNumber(userId);
     const user = { id: userId, unreadDirectsNumber };
     const payload = JSON.stringify({ user });
     await this.publisher.userUpdated(payload);
   }
 
-  async updateUnreadNotifications(userIntId) {
+  async updateUnreadNotifications(userIntId: number) {
     const [{ uid: userId }] = await dbAdapter.getUsersIdsByIntIds([userIntId]);
     const unreadNotificationsNumber = await dbAdapter.getUnreadEventsNumber(userId);
     const user = { id: userId, unreadNotificationsNumber };
@@ -45,7 +33,7 @@ export default class pubSub {
     await this.publisher.userUpdated(payload);
   }
 
-  async updateHomeFeeds(userId) {
+  async updateHomeFeeds(userId: UUID) {
     const feedObjects = await dbAdapter.getAllUserNamedFeed(userId, 'RiverOfNews');
     const payload = JSON.stringify({
       homeFeeds: feedObjects.map((f) => serializeTimeline(f)),
@@ -54,77 +42,85 @@ export default class pubSub {
     await this.publisher.userUpdated(payload);
   }
 
-  async newPost(postId) {
+  async newPost(postId: UUID) {
     const payload = JSON.stringify({ postId });
     await this.publisher.postCreated(payload);
   }
 
-  async destroyPost(postId, rooms) {
+  async destroyPost(postId: UUID, rooms: string[]) {
     const payload = JSON.stringify({ postId, rooms });
     await this.publisher.postDestroyed(payload);
   }
 
-  async updatePost(postId, rooms = null, usersBeforeIds = null) {
+  async updatePost(
+    postId: UUID,
+    rooms: string[] | null = null,
+    usersBeforeIds: UUID[] | null = null,
+  ) {
     const payload = JSON.stringify({ postId, rooms, usersBeforeIds });
     await this.publisher.postUpdated(payload);
   }
 
-  async newComment(comment) {
+  async newComment(comment: Comment) {
     const payload = JSON.stringify({ commentId: comment.id });
     await this.publisher.commentCreated(payload);
   }
 
-  async destroyComment(commentId, postId, rooms) {
+  async destroyComment(commentId: UUID, postId: UUID, rooms: string) {
     const payload = JSON.stringify({ postId, commentId, rooms });
     await this.publisher.commentDestroyed(payload);
   }
 
-  async updateComment(commentId) {
+  async updateComment(commentId: UUID) {
     const payload = JSON.stringify({ commentId });
     await this.publisher.commentUpdated(payload);
   }
 
-  async newLike(post, userId) {
+  async newLike(post: Post, userId: UUID) {
     const payload = JSON.stringify({ userId, postId: post.id });
     await this.publisher.likeAdded(payload);
   }
 
-  async removeLike(postId, userId, rooms) {
+  async removeLike(postId: UUID, userId: UUID, rooms: string[]) {
     const payload = JSON.stringify({ userId, postId, rooms });
     await this.publisher.likeRemoved(payload);
   }
 
-  async hidePost(userId, postId) {
+  async hidePost(userId: UUID, postId: UUID) {
     await this.publisher.postHidden(JSON.stringify({ userId, postId }));
   }
 
-  async unhidePost(userId, postId) {
+  async unhidePost(userId: UUID, postId: UUID) {
     await this.publisher.postUnhidden(JSON.stringify({ userId, postId }));
   }
 
-  async savePost(userId, postId) {
+  async savePost(userId: UUID, postId: UUID) {
     await this.publisher.postSaved(JSON.stringify({ userId, postId }));
   }
 
-  async unsavePost(userId, postId) {
+  async unsavePost(userId: UUID, postId: UUID) {
     await this.publisher.postUnsaved(JSON.stringify({ userId, postId }));
   }
 
-  async newCommentLike(commentId, postId, likerUUID) {
+  async newCommentLike(commentId: UUID, postId: UUID, likerUUID: UUID) {
     const payload = JSON.stringify({ commentId, postId, likerUUID });
     await this.publisher.commentLikeAdded(payload);
   }
 
-  async removeCommentLike(commentId, postId, unlikerUUID) {
+  async removeCommentLike(commentId: UUID, postId: UUID, unlikerUUID: UUID) {
     const payload = JSON.stringify({ commentId, postId, unlikerUUID });
     await this.publisher.commentLikeRemoved(payload);
   }
 
-  async globalUserUpdate(userId) {
+  async globalUserUpdate(userId: UUID) {
     await this.publisher.globalUserUpdated(JSON.stringify(userId));
   }
 
-  async updateGroupTimes(groupIds) {
+  async updateGroupTimes(groupIds: UUID[]) {
     await this.publisher.groupTimesUpdated(JSON.stringify({ groupIds }));
+  }
+
+  async newEvent(eventId: UUID) {
+    await this.publisher.eventCreated(JSON.stringify(eventId));
   }
 }