+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.93.0] - Not released
+### Added
+- The notification events are now delivered to the 'user:{userId}' realtime
+  channel with the 'event:new' message type. The message format is the same as
+  in `GET /v2/notifications` response.
+
+### Changed
+- Do not block authorization sessions if token with the invalid issue is comes.
+  We probably should add another blocking criteria in the future.
+- Remove old authorization sessions by last use (instead of last update) time.
+
 ## [1.92.1] - 2021-01-24
 ### Added
 - Count blocked auth sessions in StatsD
@@ -30,7 +41,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   'Content-Disposition: inline'. All other attachments should be served with
   'Content-Disposition: attachment' to prevent in-browser execution of active
   content.
-
 ## [1.90.1] - 2021-01-12
 ### Fixed
 - Prevent the theft of attachments from other posts when creating a post.